Start top-level API modularization
